x/cookies/types/message_bake.go
package types

import (
	sdk "github.com/cosmos/cosmos-sdk/types"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
)

const TypeMsgBake = "bake"

var _ sdk.Msg = &MsgBake{}

func NewMsgBake(creator string) *MsgBake {
	return &MsgBake{
		Creator: creator,
	}
}

func (msg *MsgBake) Route() string {
	return RouterKey
}

func (msg *MsgBake) Type() string {
	return TypeMsgBake
}

func (msg *MsgBake) GetSigners() []sdk.AccAddress {
	creator, err := sdk.AccAddressFromBech32(msg.Creator)
	if err != nil {
		panic(err)
	}
	return []sdk.AccAddress{creator}
}

func (msg *MsgBake) GetSignBytes() []byte {
	bz := ModuleCdc.MustMarshalJSON(msg)
	return sdk.MustSortJSON(bz)
}

func (msg *MsgBake) ValidateBasic() error {
	_, err := sdk.AccAddressFromBech32(msg.Creator)
	if err != nil {
		return sdkerrors.Wrapf(sdkerrors.ErrInvalidAddress, "invalid creator address (%s)", err)
	}
	return nil
}
